Add tokenizer

+++ Grammar/values.go
@@ -4,7 +4,7 @@ const (
 	// BooleanDef Default value of boolean.
 	BooleanDef bool = false
 	// IntegerDef Default value of integer.
-	IntegerDef int = 0
+	IntegerDef int8 = 0
 	// FloatDef Default value of floating types.
 	FloatDef float32 = 0
 )
